requirements.txt
accelerate==1.5.2
alpaca-py==0.39.1
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
duckduckgo-search==7.5.3
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==1.1.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
-e file:///Users/louisbrulenaudet/Documents/GitHub/agentic-market-tool
platformdirs==4.3.7
polars==1.26.0
primp==0.14.0
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
setuptools==78.0.2
six==1.17.0
smolagents==1.12.0
soupsieve==2.6
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
